Add last_modified; refactor
